hero section
after installing dependencies, updated the hero section with the hero function that returns a header with a few style divs using tailwind. We then use a slide and wrapper classnames that were defined in our index css.  we map over an array created in constants to populate a new word in the mini slideshow.